refactor: discounter Request 수정
