+++ src/main/java/christmas/domain/discount/WeekDiscounter.java
@@ -1,11 +1,10 @@
 package christmas.domain.discount;
 
-import christmas.domain.discount.dto.DiscountRequestDto;
 import christmas.domain.discount.dto.DiscountResponseDto;
-import christmas.domain.discount.dto.WeekDiscountRequestDto;
 import christmas.domain.menu.DessertMenu;
 import christmas.domain.menu.MainFoodMenu;
 import christmas.domain.menu.Menu;
+import christmas.domain.order.OrderRequestDto;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
@@ -16,12 +15,9 @@ public class WeekDiscounter implements Discounter{
     private final String WEEKEND_DISCOUNT_NAME = "주말 할인";
     private final String WEEKDAY_DISCOUNT_NAME = "평일 할인";
     @Override
-    public DiscountResponseDto discount(DiscountRequestDto discountRequestDto) {
-        if (!(discountRequestDto instanceof WeekDiscountRequestDto weekDiscountRequest)) {
-            throw new IllegalArgumentException("올바르지 않은 요청입니다.");
-        }
-        List<Menu> menus = weekDiscountRequest.getMenus();
-        LocalDate orderDate = weekDiscountRequest.getOrderDate();
+    public DiscountResponseDto discount(OrderRequestDto orderRequest) {
+        List<Menu> menus = orderRequest.getMenus();
+        LocalDate orderDate = orderRequest.getOrderDate();
 
         if (orderDate.getMonth().getValue() != 12 || orderDate.getYear() != 2023) {
             return getNoDiscount(orderDate);
