--- src/main/java/christmas/domain/discount/dto/DiscountRequestDto.java
@@ -1,4 +1,0 @@
-package christmas.domain.discount.dto;
-
-public interface DiscountRequestDto {
-}
