+++ src/main/java/christmas/domain/order/OrderRequestDto.java
@@ -1,15 +1,15 @@
-package christmas.domain.discount.dto;
+package christmas.domain.order;
 
 import christmas.domain.menu.Menu;
 
 import java.time.LocalDate;
 import java.util.List;
 
-public class FreeBieRequestDto implements DiscountRequestDto {
+public class OrderRequestDto {
     private List<Menu> menus;
     private LocalDate orderDate;
 
-    public FreeBieRequestDto(List<Menu> menus, LocalDate orderDate) {
+    public OrderRequestDto(LocalDate orderDate, List<Menu> menus) {
         this.menus = menus;
         this.orderDate = orderDate;
     }
@@ -18,7 +18,15 @@ public class FreeBieRequestDto implements DiscountRequestDto {
         return menus;
     }
 
+    public void setMenus(List<Menu> menus) {
+        this.menus = menus;
+    }
+
     public LocalDate getOrderDate() {
         return orderDate;
+    }
+
+    public void setOrderDate(LocalDate orderDate) {
+        this.orderDate = orderDate;
     }
 }
