--- src/test/java/christmas/domain/discount/InvalidTestDto.java
@@ -1,17 +1,0 @@
-package christmas.domain.discount;
-
-import christmas.domain.discount.dto.DiscountRequestDto;
-
-import java.time.LocalDate;
-
-public class InvalidTestDto implements DiscountRequestDto {
-    private LocalDate orderDate;
-
-    public InvalidTestDto(LocalDate orderDate) {
-        this.orderDate = orderDate;
-    }
-
-    public LocalDate getOrderDate() {
-        return orderDate;
-    }
-}
